Split s3stalker and daemon parts
