+++ bin/s3stalker.py
@@ -54,7 +54,6 @@ def main():
 
     bucket = args.bucket
     config = read_yaml(args.config)
-    time_delta = config['timedelta']
 
     if args.log is not None:
         with open(args.log) as fd:
@@ -62,7 +61,7 @@ def main():
             logging.config.dictConfig(log_dict)
 
     try:
-        publish_new_files(bucket, config, time_delta)
+        publish_new_files(bucket, config)
     except KeyboardInterrupt:
         print("terminating publisher...")
