+++ pytroll_collectors/s3stalker.py
@@ -19,111 +19,17 @@
 
 import logging
 import posixpath
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timedelta
 import time
 from dateutil import tz
 from contextlib import contextmanager
 import s3fs
 from posttroll.publisher import Publish
-from posttroll.publisher import create_publisher_from_dict_config
 from trollsift import Parser
-
-import signal
-from threading import Thread
 
 from pytroll_collectors.fsspec_to_message import filelist_unzip_to_messages
 
 logger = logging.getLogger(__name__)
-
-
-class S3StalkerRunner(Thread):
-    """Runner for stalking for new files in an S3 object store."""
-
-    def __init__(self, bucket, config, startup_timedelta_seconds):
-        """Initialize the S3Stalker runner class."""
-        super().__init__()
-
-        self.bucket = bucket
-        self.config = config
-        self.startup_timedelta_seconds = startup_timedelta_seconds
-        self.time_back = self.config['timedelta']
-        self._timedelta = self.config['timedelta']
-        self._wait_seconds = timedelta(**self.time_back).total_seconds()
-
-        self.publisher = None
-        self.loop = False
-        self._set_signal_shutdown()
-
-    def _set_signal_shutdown(self):
-        """Set a signal to handle shutdown."""
-        signal.signal(signal.SIGTERM, self.signal_shutdown)
-
-    def _setup_and_start_communication(self):
-        """Set up the Posttroll communication and start the publisher."""
-        self.publisher = create_publisher_from_dict_config(self.config['publisher'])
-        self.publisher.start()
-        self.loop = True
-
-    def signal_shutdown(self, *args, **kwargs):
-        """Shutdown the S3 Stalker daemon/runner."""
-        self.close()
-
-    def run(self):
-        """Start the s3-stalker daemon/runner in a thread."""
-        logger.info("Starting up s3stalker.")
-        self._setup_and_start_communication()
-
-        first_run = True
-        last_fetch_time = None
-        while self.loop:
-            self._set_timedelta(last_fetch_time, first_run)
-
-            last_fetch_time = get_last_fetch()
-            logger.debug("Last fetch time...: %s", str(last_fetch_time))
-            first_run = False
-
-            self._process_messages()
-
-            logger.debug("Waiting %d seconds", self._wait_seconds)
-            time.sleep(max(self._wait_seconds, 0))
-
-    def _set_timedelta(self, last_fetch_time, first_run):
-        self._timedelta = self._get_timedelta(last_fetch_time, is_first_run=first_run)
-
-    def _process_messages(self):
-        """Go through all messages in list and publish them one after the other."""
-        messages = create_messages_for_recent_files(self.bucket, self.config, self._timedelta)
-        for message in messages:
-            logger.info("Publishing %s", str(message))
-            self.publisher.send(str(message))
-
-    def _get_timedelta(self, last_fetch_time, is_first_run):
-        """Get the seconds for the time window to search for (new) files."""
-        if is_first_run:
-            logger.info('Create messages with urls for most recent files only')
-            logger.info('On start up we consider files with age up to %d seconds from now',
-                        self.startup_timedelta_seconds)
-            return {'seconds': self.startup_timedelta_seconds}
-
-        seconds_back = self._get_seconds_back_to_search(last_fetch_time)
-        return {'seconds': seconds_back}
-
-    def _get_seconds_back_to_search(self, last_fetch_time):
-        """Update the time to look back considering also the modification time of the last file."""
-        if last_fetch_time is None:
-            return self._wait_seconds
-
-        start_time = datetime.utcnow()
-        start_time = start_time.replace(tzinfo=timezone.utc)
-        seconds_to_last_file = (start_time - last_fetch_time).total_seconds()
-        return max(self._wait_seconds, seconds_to_last_file)
-
-    def close(self):
-        """Shutdown the S3Stalker runner."""
-        logger.info('Terminating the S3 Stalker daemon/runner.')
-        self.loop = False
-        if self.publisher:
-            self.publisher.stop()
 
 
 @contextmanager
@@ -140,7 +46,7 @@ def sleeper(duration):
 class DatetimeHolder:
     """Holder for the last_fetch datetime."""
 
-    last_fetch = datetime.utcnow() - timedelta(hours=12)
+    last_fetch = datetime.now(tz.UTC) - timedelta(hours=12)
 
 
 def get_last_files(path, *args, pattern=None, **kwargs):
@@ -161,6 +67,7 @@ def _reset_last_fetch_from_file_list(files):
 
 def _get_files_since_last_fetch(fs, path):
     files = fs.ls(path, detail=True)
+    logger.warning(str(files))
     files = list(filter((lambda x: x['LastModified'] > DatetimeHolder.last_fetch), files))
     return files
 
@@ -190,26 +97,26 @@ def get_last_fetch():
     return DatetimeHolder.last_fetch
 
 
+def publish_new_files(bucket, config):
+    """Publish files newly arrived in bucket."""
+    time_back = config['timedelta']
+    with Publish("s3_stalker") as pub:
+        with sleeper(2.5):
+            messages = create_messages_for_recent_files(bucket, config, time_back)
+        for message in messages:
+            logger.info("Publishing %s", str(message))
+            pub.send(str(message))
+
+
 def create_messages_for_recent_files(bucket, config, time_back):
     """Create messages for recent files and return."""
     logger.debug("Create messages for recent files...")
     logger.debug("time_back = %s", str(time_back))
     subject = config['subject']
     pattern = config.get('file_pattern')
-    with sleeper(2.5):
-        set_last_fetch(datetime.now(tz.UTC) - timedelta(**time_back))
-        s3_kwargs = config['s3_kwargs']
-        fs_, files = get_last_files(bucket, pattern=pattern, **s3_kwargs)
-        messages = filelist_unzip_to_messages(fs_, files, subject)
-        # fs_.clear_instance_cache()
+    set_last_fetch(datetime.now(tz.UTC) - timedelta(**time_back))
+    s3_kwargs = config['s3_kwargs']
+    fs_, files = get_last_files(bucket, pattern=pattern, **s3_kwargs)
+    messages = filelist_unzip_to_messages(fs_, files, subject)
 
     return messages
-
-
-def publish_new_files(bucket, config, time_back):
-    """Publish files newly arrived in bucket."""
-    messages = create_messages_for_recent_files(bucket, config, time_back)
-    with Publish("s3_stalker") as pub:
-        for message in messages:
-            logger.info("Publishing %s", str(message))
-            pub.send(str(message))
